COM-3257 - rename file
